+++ R/app_ui.R
@@ -82,6 +82,7 @@ create_sidebar <- function() {
 #' Creates app's body
 create_body <- function() {
   bs4Dash::bs4DashBody(
+    fresh::use_theme(theme_bs4Dash()),
     bs4Dash::bs4TabItems(
       bs4Dash::bs4TabItem(
         tabName = "summary",